requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.27.9
cryptography==41.0.7
distlib==0.3.4
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.2.1
filelock==3.6.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
lxml==5.2.2
outcome==1.3.0.post0
packaging==23.2
platformdirs==2.5.2
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.14.1
websocket-client==1.8.0
wsproto==1.2.0
